+++ src/bedrock/symbols/linux.h
@@ -11,7 +11,7 @@
 
 namespace endstone::runtime {
 
-static constexpr std::array<std::pair<std::string_view, std::size_t>, 72> symbols = {{
+static constexpr std::array<std::pair<std::string_view, std::size_t>, 74> symbols = {{
     {"BlockState::StateListNode::mHead", 245468000},
     {"Enchant::mEnchants", 245294016},
     {"MobEffect::mMobEffects", 245274360},
@@ -103,6 +103,8 @@ static constexpr std::array<std::pair<std::string_view, std::size_t>, 72> symbol
     {"_ZN6Player14stopSleepInBedEbb", 188148304},
     {"_ZN6Player15startSleepInBedERK8BlockPosbf", 188144032},
     {"_ZN6Player17completeUsingItemEv", 188167680},
+    {"_ZN6Player18setRespawnPositionERK8BlockPos13DimensionType", 188285872},
+    {"_ZN6Player28setSpawnBlockRespawnPositionERK8BlockPos13DimensionType", 188285152},
     {"_ZN6Player4dropERK9ItemStackb", 188279872},
     {"_ZN6Player4takeER5Actorii", 188227136},
     {"_ZN6Player9addLevelsEi", 188298880},